Avoid merging 'optionalDependencies' into 'dependencies'
Optional dependencies should not be merged into the list of dependencies
for the same reason that development dependencies shouldn't be merged
into the list of dependencies: they need to be treated differently.

Merging these objects in normalize-package-data bubbles up this bug
through read-package-json and into init-package-json, which creates a
problem where `npm init -y` will copy optional dependencies into the
`dependencies` object.

Instead of normalizing the dependency metadata, this module is mutating
it to have a different meaning, which seems to be a bug. This commit
resolves the mutation and treats the `optionalDependencies` key the same
way as the `devDependencies` key.

Fixes: https://github.com/npm/cli/issues/724